Move NodeAccessor into test_utils
